♻️ Move sync types to transcribee_proto
