+++ backend/openapi-schema.yml
@@ -493,6 +493,7 @@ components:
           type: string
           format: date-time
           readOnly: true
+          nullable: true
         completed_at:
           type: string
           format: date-time
